Localisation of 'Alert' state and ability to name sensors (#13)
This PR proposes to add Swedish localisation of the 'Alert' state in the same way as 'No Alerts' (no other locales supported currently). It also suggests to add ability to name the sensor, in case you want several. For example, living near the coast, you may want both alerts for sea as well as land. And you want naming to be deterministic.

As as side note, we should really split out the lovelace card and try to get this into HACS, for maintainability. And maybe to get support into the Meteoalarm Card as well.